Refactor: decimal para montos y tests unitarios de transacciones
- Se reemplaza double por decimal en entidades, DTOs y vistas para mayor precisión en montos.
- Mejora la lógica de actualización de transacciones considerando el monto y tipo de operación anterior.
- Métodos de obtención por ID ahora retornan tipos anulables.
- Validaciones y mensajes de error más claros y robustos.
- Ajustes en formularios para soportar decimal y type="number".
- Se agrega proyecto de pruebas unitarias para TransactionService con xUnit, Moq y FluentAssertions.
- Actualización de interfaces y solución para reflejar estos cambios.

+++ src/BudgetManager.Web/Models/Account/AccountListVM.cs
@@ -6,7 +6,7 @@ public class AccountListVM
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public double Balance { get; set; } = 0.0;
+    public decimal Balance { get; set; } = 0.0m;
     public int AccountTypeId { get; set; }
     public string AccountType { get; set; }
     public string? Description { get; set; }
